Change ParameterDelegate's generic type semantics
Leave the class's generic type as unused, so it's easy to justify casting an instance to something different.

Instead, give methods their own generic type that's the same as the property passed in, the property can be checked against the declared property, and argument types can be reasoned about based on the property type.