Improve lazy argument `parameter()` performance by double
Also add tests and re-organize the `ParameterizeScope` methods so they appear in the order they're called in the DSL.

Inline everything to save on unnecessary instantiations, and don't use the stdlib `lazy` since it has a decent amount of unnecessary thread safety overhead (and can't be inlined).

Also tried using the stdlib `Iterable {}` function, and storing the calculated arguments in a function variable, but that was a tad slower (maybe due to it being wrapped into a new reference object instance as it's captured into `Iterable()`'s closure).