Add `ParameterizeFailedError`, and throw it in the default `throwHandler`
The default throw handler no longer `print`s anything (which feels like a side effect libraries shouldn't have by default). Now the default `throwHandler` wraps the failure in a `ParameterizeFailedError`, with the same info written into its error message. The outer error's stack trace is also cleared, so it's essentially the failure's stack trace, but with a list of parameter arguments preceding it.

It's an `Error` instead of an `Exception` since the failure being wrapped could itself be an `Error`. `Error`s have "should not be caught" semantics, so wrapping the failure in a hypothetical `ParameterizeFailedException` instead would hide the failure `Error`'s catching semantics. That, and also because in its motivating use case of testing, it isn't meant to be caught (similar to how `AssertionError` is used by kotlin.test). In the case that a library consumer *does* need it to be an `Exception`, it is possible to fully customize the `throwHandler`, even re-implementing `ParameterizeFailedException` as an `Exception` instead.