Make `parameterize` configuration usage contextual
A context-less overload was also added so `parameterize` can be run with a default ParameterizeContext. Since this overload is likely to be the first used by library users, its brief docs point to the primary contextual `parameterize` where all the detailed docs are, and also to the `ParameterizeContext` interface which guides through setting up and pulling in a configuration.

Making it contextual also helps avoid using global mutable state in the API with `ParameterizeConfig.default`.