Move `ParameterDelegate` to `ParameterizeScope.ParameterDelegate`
The `ParameterDelegate` class is only used for the `parameterize` DSL, so nesting it in `ParameterizeScope` better indicates its purpose, and removes a likely common class name from the global namespace.

Originally kept in a separate file since it was filled with complicated logic, so kept the old logic in place renaming it to `ParameterState`, and have the new `ParameterDelegate` be composed of it.