Simplify implementation for the new argument behavior
Now that arguments are calculated immediately, there's no intermediate parameter state before initializing. This means a lot of the parameter tracking and argument iteration logic could be simplified. Documentation was also updated to reflect the changes in how the parameter state is represented.

Tests also no longer need to use a parameter to ensure that it's iterated over in the combinations, so a lot of `useParameter` lines could be removed as well.

Additionally, tests that only specify how used/unused parameters behave have been removed, since they are no longer cases that change the behavior, so they don't need to be considered.