Improve failure tooling on jvm with opentest4j
Resolves #13

`ParameterizeFailedError` now extends `MultipleFailuresError`, providing a list of failures that tools can hook into. The listed errors are not decorated with additional parameter argument info, so tools can see the raw errors that were thrown and handle them appropriately.

On IntelliJ, for example, this means each failure from `assertEquals()` and the like will have a "<Click to see difference>" link that pulls up a diff UI for the actual/expected values.