Clean up `ParameterizeException` checks
This moves the `ParameterizeException` class is now in its own file, and adds has a more concise `checkState` function that simplifies validation code, pulling in the `ParameterizeState` contextually. Similar to Kotlin's `check` function, now read as "make sure value is correct" instead of a negated "if value isn't correct, throw".