Remove `ParameterizeConfiguration.` prefix from scope classes in code/docs
Notably in the KDoc, having ParameterizeConfiguration spelled out on every line of `parameterize` took up a lot of horizontal space. Especially on mobile, where even on a larger phone each parameter definition (name, receiver lambda type, default argument) could span 4 lines.

There doesn't appear to be a way to configure a display name through Dokka, so a small snippet was added to the Gradle build to replace the prefix in the html files after they're generated.