Change `parameterize` to be inline
resolves #20

This enables coroutines from outside the `parameterize` call to work through the block. In addition to the other loop-like benefits that come with `inline`, like returning from outer functions and cleaner stack traces.

The main looping logic is now factored out, exposing a minimal API appropriate for inlining. Now the function itself can be a simple loop with one call for iterating, and another to signal failures.

The `decorator` also had to be changed to accommodate this. The `decorator` isn't inlinable since it could be pulled from the configuration. Because the `parameterize` block is wrapped by the `decorator`, block would have to be `crossinline`, but coroutines can't work through `crossinline` lambdas. To solve this, the `decorator` was changed to be a `suspend` lambda so its execution could be broken into two parts. The "before iteration" logic gets executed, then the `decorator` suspends while the inlined `parameterize` block is executed, and finally it's resumed again for the "after iteration" logic.

Because the `decorator` configuration option is a fairly significant binary change, and providing binary compatibility stubs isn't straightforward. The `parameterize` function itself would be easy enough to provide a new stub for, but stubs for the `decorator` property in ParameterizeConfiguration and its builder, not so much. Because of this, the next release will drop backwards compatibility with v0.2 and earlier, and the binary compatibility stubs in Parameterize.deprecated.kt will be removed since they're no longer useful.