Let the `decorator` finish before invoking `onFailure`
Having `onFailure` run *after* the `decorator` is easier to reason about if the failure handler itself throws. The `decorator` has "always completes" semantics, and `onFailure` has "terminates immediately" semantics, and that's only possible if `onFailure` runs after. If `onFailure` were to run immediately for a failure in the parameterize block, then the `decorator`'s semantics with how it interacts with failures wouldn't be so straightforward.