Change default `onFailure` to re-throw the failure
Resolves #21

Changing `parameterize` to re-throw immediately means it behaves the same as any other loop by default, immediately breaking without any special handling. This also means that `parameterize` no longer pushes forward to the next iteration, throwing a collective `Error` by default, which is only really desirable when used for testing since `Error`s aren't meant to be caught.

A subtle advantage is that `onFailure` doesn't prescribe a default number of failures to record, and that choice can be offloaded onto the users of the library to decide, instead of imposing "only the first 10"/"record them all"/etc.