Expose the `parameterize` overload that takes a configuration
Useful for library users to be able to call this directly with a pre-built configuration, instead of building a new one every call.

Notably for specialized `parameterize`-wrapping functions that will have their own configuration that won't be changed, like a parameterless `parameterizeTest` function that would record failures and hook into the test lifecycle with decorators.