Remove contextual `parameterize` overload
Resolves #22

Having a contextual overload is proving to be problematic (both because it's leaning into a future language feature, for which the prototype is constantly updating. And because contextual overloads are specifically discouraged by the proposal).

Instead, having devs wrap `parameterize` in their own pre-configured function is proving to be a much more powerful and misuse-resistant approach.