Change `parameterize` signatures to be more flexible and performant
Now that a configuration instance is being passed around no matter what, it makes sense to make the configuration-only `parameterize` be the primary overload, avoiding a configuration being created every argumentless call to `parameterize {}`.

The configuration-only overload's `configuration` now has a default argument, being set to a globally shared default configuration instance if one isn't passed. This works nicely, since even though both `parameterize` overloads can be called with all default options, Kotlin chooses the overload with the fewest options, which in this case is the configuration-only overload that doesn't instantiate a new configuration.

Additionally, to make the options overload more flexible, and to bring back the "tweak existing configuration" functionality that was lost with the removal of the contextual overload, a `configuration` parameter was added. This makes the signature more symmetrical with the configuration-only overload, and without any added runtime cost since a configuration is being copied either way, just with more control over where it's copied from now. Additionally, since the configuration parameter was added at the start, it has the added benefit of encouraging named arguments to be used.