Enable the `parameter()` API to be extended
The `Parameter.arguments` property has been changed to be public, that way it's possible to operate on a parameter after it's been created.

Additionally, the `Parameter` constructor has been made public, making it possible to create a new `Parameter` without needing to be in the `ParameterizeScope`. Without this, it's not possible to write an extension function on `Parameter` that creates a new one, unless the scope is explicitly taken in as another argument so `parameter()` can be called. In the future this won't be necessary once context parameters are introduced to Kotlin, but for now, this makes for a good stopgap. Once context parameters *have* been introduced, this constructor will be deprecated in favor of `parameter()`.