Change Parameterize license to Apache 2.0
Originally the GPL 3.0 license was decided on so that any improvements or derivative projects that benefit from this project would also have to their source be shared and open.

However, GPL imposes itself even if this library is only being used as a dependency. LGPL would've been more in line with the original intent for the license, since it more so applies to modification of the project/binaries, and doesn't impose itself on other projects when used as a dependency. But even so, though, there's concern about licensing under any of the GPL-based licenses because it's widely considered to be high risk in professional settings. The potential for litigation if accidentally misused means the GPL licenses won't be as likely to be whitelisted for use in corporate projects, and that's concerning for Parameterize's adoption.

The Apache 2.0 license was chosen instead because of this concern since it's more permissive, and common on license whitelists because of it being low risk. It was chosen over other permissive licenses (e.g. the MIT license) as it seems to be the license of choice in the Kotlin ecosystem, being used by all the first-party Kotlin libraries, and a comfortable majority overall.