Add a `@ParameterizeDsl` marker to discourage nesting parameters
The `@ParameterizeDsl` annotation was made public instead of internal since there are reasonable use cases for extending the parameterize syntax, and being able to create custom functions with DSL-restricting scopes is an important part of that.

Declaring a parameter within another parameter's `iterator()` call (e.g. within a lazy `parameter {}` block) is already an error, but only at runtime. Now with `@ParameterizeDsl`, it won't be possible to declare nested parameters with the normal syntax. It *is* still possible, but with an uglier explicit `this` syntax that's harder to accidentally write, meaning devs would have to go out of their way to use it incorrectly.

Since `LazyParameterScope` only serves as a compile-time check, it is an inline value class to avoid any runtime overhead from instantiating an instance that won't be used. The `LazyParameterScope` is backed by the `ParameterizeScope`, so if any logic does need to be added later on, the `LazyParameterScope` constructor will be able to tap into the current `parameterize` state.

The `LazyParameterScope` class was also declared at the top level (unlike `Parameter` and `ParameterDelegate`) because it's really an extension of the `ParameterScope`, and it's not directly part of the `ParameterizeScope` syntax, just the top-level lazy `parameter {}` function. The fully qualified `ParameterizeScope.LazyParameterScope` name, especially as part of the lambda signature, is also undesirably long in code and in documentation. Ideally the `Parameter` and delegate classes would've also be top level, but being such a common word in programming, `Parameter` could too easily conflict with another class of the same name from outside this library.