Add Wasm targets
Resolves #28

Minimally implement wasm-js and wasm-wasi targets for now without stack
trace clearing, since they're still experimental and APIs might change
in the future adding a maintenance burden.

For now they're not a shared source set, but can be changed to be one in
the future when they're stable if they have the same capabilities and do
share the same implementation.