Fix property comparison for Wasm targets
Previously, comparing KProperties from delegates would either work
completely (jvm/native), or not at all (js), so a quick runtime check
was enough to determine whether KProperty.equals could be used. But now
the Wasm targets only sometimes work, and other times don't, depending
on the property and how it's declared, which resulted in some usages
failing. To properly solve this, the `KProperty.equalsProperty` function
was made into an `expect` function, and each platform actualizes it
according to whether `KProperty.equals` works or not.