Fix `ParameterizeFailedError` count including skipped iterations
When a parameter is declared with empty arguments causing the iteration
to be skipped, that iteration would still count towards the total when
reported in a `ParameterizeFailedError` despite it not finishing. When
used for writing tests, this means it would look like some cases didn't
fail, since the "fail/total" result wouldn't look like 100% failures.

This change also expose the number of successful/skipped iterations in
the `onComplete` API now that they're being tracked.