Change ParameterizeScope into an interface
This allows 3rd party code to hook into parameter declarations/reads,
making it possible to implement additional behavior/tracking.

This also makes it possible to combine with other DSLs by implementing
the scope interfaces for both in a single scope class.

The interface and `parameter()` functions are moved to their own file
since they are roughly self-contained scope functionality. This file may
also be moved to a separate `core` library in the future, so having
them all be in the same class means they can be moved together in a
binary-compatible way e.g. with the top-level functions still being in
the `ParameterizeScopeKt` class on JVM.