Remove `@ParameterizeDsl` in favor of deprecated overload
Having `@ParameterizeDsl` for lazy `parameter {}` functions was too
restrictive, since it prohibited using parameter functions *at all*,
when really all that should be restricted is their declaration. There
are valid use cases for calling `parameter()` functions (and
`ParameterizeScope` extensions in general) in lazy parameters, such as
including/transforming pre-defined parameters, so permitting them to be
used would be preferable.

Instead, a less restrictive option would be to overload the
`provideDelegate` operator used for parameter declaration. Parameters
declared in the lazy parameter block will now call an error-deprecated
overload with the same signature, dissuading declaration in the lazy
block. This can be worked around, but so could the DSL marker, so there
isn't any compromise with this approach.