Use `DeclaredParameter` instead of `ParameterizeFailure.Argument`
The two classes filled the same role, introspecting parameters that have
been declared, and especially now that the Parameter delegate is named
"DeclaredParameter", this is a natural change.

`ParameterizeFailure` had its `toString` removed, since `Throwable`s
already don't typically have meaningful `toString` implementations,
and now `parameters` doesn't have a very meaningful representation since
it would render as a list of values without parameters.

Property validation for parameters was moved from `ParameterState` to
`ParameterizeState` since `getDeclaredParameter` is now used in multiple
places, and not necessarily requiring validation. It is also more a
responsibility of the `parameterize` DSL as well, so it makes sense this
way.