Account for non-local returns when ending iterations
Declaring a parameter with an iteration's scope after leaving that
iteration shouldn't be possible, and there are checks for most cases,
but early returning out of the block was not accounted for when checking
if the iteration ended.

The `decorator` code also does not complete when returning non-locally,
even though it always should.