Fix KDoc not picking up the version from CI
The `CI_VERSION` environment variable was picked up and set from within
publishing-conventions.gradle.kts, but that isn't included in the root
project where the KDoc is generated. Because of that, Dokka used the
snapshot version that's in the Gradle properties.

Instead of having build scripts pick up and set the version, now the CI
sets the project version directly using the `ORG_GRADLE_PROJECT_version`
environment variable, overriding the version from the Gradle properties.